requirements.txt
asgiref
channels
dwebsocket
jieba
whoosh
django-haystack
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-allauth==0.39.1
django-appconf==1.0.3
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-imagekit==4.0.2
django-import-export==1.2.0
django-js-asset==1.2.2
django-model-utils==3.1.2
django-mptt==0.10.0
django-notifications-hq==1.5.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
idna==2.8
jdcal==1.4
jsonfield==2.0.2
mysqlclient==1.4.2.post1
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
pilkit==2.0
Pillow==6.0.0
pycodestyle==2.5.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.24.2
xlrd==1.2.0
xlwt==1.3.0
